requirements.txt
lcu_driver==3.0.1
Pillow==10.4.0
tqdm==4.66.2